Ignore var-naming[pattern] to foreign role vars (#4095)
The changes to the utils.py's _rolepath were to enable the function to
handle fully qualified collection names (FQCNs). Otherwise, the function
would be unable to parse the role name from the FQCN returning basedir
as the role_path. This can result in additional files with violations
being returned by the _look_for_role_files function.