Treat the default inventory_files as host list (#4448)
The --inventory option to the ansible-playbook executable can accept a
value that is comma-separated. This comma-separated value is presumed to
be a host list and not a list of inventory files. This was the behavior
for the inventory_files variable before 6fbb12e9. There wasn't a reason
for its removal and it was an oversight made.